requirements.txt
python-telegram-bot>=20.0
pytest>=7.0
loguru>=0.7
pre-commit>=3.6
python-dotenv>=1.0.0

